Refactors the constraints
